+++ config/schc.h
@@ -15,9 +15,9 @@
 #define MAX_NB_RULE_IDS 17
 
 extern char FR_MODE[];
-extern int RETRANSMISSION_TIMEOUT;
-extern int SIGFOX_DL_TIMEOUT;
-extern int INACTIVITY_TIMEOUT;
+extern float RETRANSMISSION_TIMEOUT;
+extern float SIGFOX_DL_TIMEOUT;
+extern float INACTIVITY_TIMEOUT;
 extern int MAX_ACK_REQUESTS;
 extern const int L2_WORD_SIZE;
 extern int DELAY_BETWEEN_FRAGMENTS;
